Extracts Data Source from CSV class
